+++ Runtime/Sets/BoolSet.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UnderLogic.Collections
+namespace UnderLogic.Collections.Sets
 {
     [CreateAssetMenu(menuName = "Collections/Sets/Bool Set")]
     public class BoolSet : RuntimeSet<bool> { }