+++ Runtime/Sets/ColorSet.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UnderLogic.Collections
+namespace UnderLogic.Collections.Sets
 {
     [CreateAssetMenu(menuName = "Collections/Sets/Color Set")]
     public class ColorSet : RuntimeSet<Color> { }