+++ Runtime/Sets/IntSet.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UnderLogic.Collections
+namespace UnderLogic.Collections.Sets
 {
     [CreateAssetMenu(menuName = "Collections/Sets/Int Set")]
     public class IntSet : RuntimeSet<int> { }