+++ Runtime/Sets/ScriptableObjectSet.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UnderLogic.Collections
+namespace UnderLogic.Collections.Sets
 {
     [CreateAssetMenu(menuName = "Collections/Sets/ScriptableObject Set")]
     public class ScriptableObjectSet : RuntimeSet<ScriptableObject> { }