+++ Runtime/Sets/TransformSet.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UnderLogic.Collections
+namespace UnderLogic.Collections.Sets
 {
     [CreateAssetMenu(menuName = "Collections/Sets/Transform Set")]
     public class TransformSet : RuntimeSet<Transform> { }