+++ Runtime/Sets/Vector2IntSet.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UnderLogic.Collections
+namespace UnderLogic.Collections.Sets
 {
     [CreateAssetMenu(menuName = "Collections/Sets/Vector2Int Set")]
     public class Vector2IntSet : RuntimeSet<Vector2Int> { }