+++ Runtime/Sets/Vector3Set.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UnderLogic.Collections
+namespace UnderLogic.Collections.Sets
 {
     [CreateAssetMenu(menuName = "Collections/Sets/Vector3 Set")]
     public class Vector3Set : RuntimeSet<Vector3>