+++ Runtime/Sets/Watchers/Vector3SetWatcher.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UnderLogic.Collections.Watchers
+namespace UnderLogic.Collections.Sets.Watchers
 {
     public class Vector3SetWatcher: RuntimeSetWatcher<Vector3, Vector3Set> { }
 }