+++ Samples~/SetWatchers/Scripts/SelectionController.cs
@@ -1,3 +1,4 @@
+using UnderLogic.Collections.Sets;
 using UnityEngine;
 
 namespace UnderLogic.Collections.Samples