+++ Samples~/Sets/Scripts/NameListDisplay.cs
@@ -1,3 +1,4 @@
+using UnderLogic.Collections.Sets;
 using UnityEngine;
 using UnityEngine.UI;