+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2022-12-30
+
+## Changed
+
+- Moved all `RuntimeSet<T>` classes to `UnderLogic.Collections.Sets` namespace
+- Moved all `RuntimeSetWatcher<T, TSet>` classes to `UnderLogic.Collections.Sets.Watchers` namespace
+
 ## [0.2.3] - 2022-12-29
 
 ### Added